Lock no Python

+++ scripts/08-python/install.sh
@@ -2,4 +2,9 @@
 
 echo " - Python..."
 
+LOCK=/tmp/workstation.lock.python
+[ -f ${LOCK} ] && return
+
 apt-get install -y python3 > /dev/null
+
+> ${LOCK}
